Move NavBar component to components folder (#29)
