+++ frontend-service/src/pages/Questions.js
@@ -1,7 +1,8 @@
 import axios from "axios";
 import { useState, useEffect } from "react";
 
-import NavBar from "./components/NavBar";
+import NavBar from "../components/NavBar";
+
 import Paper from "@mui/material/Paper";
 import Table from "@mui/material/Table";
 import TableBody from "@mui/material/TableBody";
